Fix speed change bug and remove unnecessary javascript
-Previously, the Youtube API had a bug where a speed change from
another speed back to 1.0 would not register properly in Firefox.
The API has since been fixed, so there is no need for a workaround

-There was previously a bug in the Youtube API that caused videos
to not play at a set speed in Safari. The bug has apparently since
been fixed, so we no longer need code for that browser
special case.

-Also modifies tests to reflect changes in video player javascript