Fixes issue with LTI consumer error with adding vertical blocks
While it was possible to embed sequentials and html blocks in consumers'
courses, the server would previously crash if an LTI consumer tried to
add a vertical to their page. This is because, when retrieving a vertical,
the server expected certain keys to exist in dicts; they wouldn't
if the request was made by an LTI consumer.