Fix wonky tests from hinting
This was introduced back in 2014 as part of Nick Parlante's Custom
Hinting/Feedback work.

We've run into some issues with this over the years.
- It was PR'ed upstream, but we pre-emptively cherry-picked it back to
  our remote before waiting for an upstream merge.
- As a consequence, we must have resolved some conflicts incorrectly or
  ran into some other issues.

This referred to a non-existent tab when adding new components.
- Add New Component
- Problem
- See: Common Problem Types and Advanced

Since this 3rd tab hasn't existed for more than a year,
I don't know what it was doing :\