Unfork course rerun tests
While edX enables course reruns by default, we disable it on all of our
instances. Instead of forking the tests to accomodate, we should instead
enable the feature during tests.  The added benefit here is that this
ensures that we continue testing the feature, even though we aren't
(currently) using it.