Add tests for extra fields for `affiliation`
When `affiliation` is selected to be "Stanford University,
`stanford_department` and `sunet_id` is required. Function
`clean_affiliation` in `forms.py` checks this. Previously fixed bug
regarding frozen registration screen on invalid inputs (PR #650) and
discovered missing unit tests. Added tests check that form is invalid
given sunet_id that is missing or not 3-8 chars long, or
`stanford_department` that is missing or invalid.