Fix MetaData Editor failing tests
Previously, running JS tests results in four failing tests, `Test
Metadata Editor MetadataView.Editor creates editors for each field`, as
a result of attempting to access properties of a null `options` and
triggering a `TyperError`. This fix adds in a check to only access
`options.warning` when `options` is non-null.