Fix TypeError: undefined editorView on hide
Previously, the unit test "editing an xblock: can show an edit modal for a
child xblock with broken JavaScript" would fail. This was because when
an edit modal was created and shown, the `hide` function would be enabled
before `displayXBlock` was called. This meant the `editorView` could be
accessed in `hide` before it was ever initialized, which would result in a
`TypeError: this.editorView is undefined`. This fix rearranges the order
so that `editorView` is initialized first.