Accept valid CME sub-specialties
While the model doesn't enforce choices for the sub-specialty,
the form apparently does. We had been arbitrarily defaulting the
sub-specialty input dropdown with only the blank entry `-----`, since we
overwrite after the specialty is updated. A side-effect of this was that
the form backend thought any non-empty sub-specialty was invalid.

Valid sub-specialty values are now allowed.