Unfork unused SASS
This appears to have been introduced (mistakenly?) during conflict
resolution performed by jbau back in 2013! [1][2]

With this change, this SASS file no longer contains any Stanford
modifications.

There are currently no references to this rule in `edx-platform`.
It is, however, referenced in our themed file by the same name.

This raises the larger issue that this leads me to believe that due to
COMP_THEMING, our old copy of this file on our theme is shadowing the
platform version. This means, not only are we carrying technical debt in
the form of unused styling, but worse, we haven't been receiving styling
updates from edx.

[1] 1cedb45b66f80e2dfa6d9f0ac08bbc18e1189f36
[2] https://github.com/Stanford-Online/edx-platform/commit/1cedb45b66f80e2dfa6d9f0ac08bbc18e1189f36